refactor: translations, add error boundary, and update app branding
