+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2022 He Linming
+Copyright (c) 2022–2026 He Linming
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
