+++ src/components/error-boundary.tsx
@@ -1,0 +1,40 @@
+import { Component } from "react"
+
+interface Props {
+  children: React.ReactNode
+  fallback?: React.ReactNode
+}
+
+interface State {
+  error: Error | null
+}
+
+export class ErrorBoundary extends Component<Props, State> {
+  constructor(props: Props) {
+    super(props)
+    this.state = { error: null }
+  }
+
+  static getDerivedStateFromError(error: Error): State {
+    return { error }
+  }
+
+  render() {
+    if (this.state.error) {
+      return (
+        this.props.fallback ?? (
+          <main className="container mx-auto p-4 pt-16">
+            <h1 className="mb-2 text-2xl font-semibold">
+              Something went wrong
+            </h1>
+            <p className="text-muted-foreground">
+              An unexpected error occurred. Please try refreshing the page.
+            </p>
+          </main>
+        )
+      )
+    }
+
+    return this.props.children
+  }
+}
