--- src/hooks/use-theme.ts
@@ -1,1 +1,0 @@
-export { useTheme } from "@/components/theme-provider"
